+++ src/ui/popup/components/body.tsx
@@ -14,10 +14,12 @@ import {getDuration} from '../../../utils/time';
 import {DONATE_URL, HOMEPAGE_URL, MOBILE_URL, getHelpURL} from '../../../utils/links';
 import {getLocalMessage} from '../../../utils/locales';
 import {compose, openExtensionPage} from '../../utils';
+import {PlusBody} from '@plus/popup/plus-body'; // eslint-disable-line
 import type {ExtensionData, ExtensionActions, News as NewsObject} from '../../../definitions';
 import {isMobile} from '../../../utils/platform';
 
 declare const __THUNDERBIRD__: boolean;
+declare const __PLUS__: boolean;
 
 interface BodyProps {
     data: ExtensionData;
@@ -54,6 +56,10 @@ function Body(props: BodyProps & {fonts: string[]}) {
                 <Loader complete={false} />
             </body>
         );
+    }
+
+    if (props.data.settings.previewNewestDesign && __PLUS__) {
+        return <PlusBody {...props} fonts={props.fonts} />;
     }
 
     if (isMobile || props.data.settings.previewNewDesign) {