+++ src/utils/validation.ts
@@ -178,6 +178,7 @@ export function validateSettings(settings: Partial<UserSettings>): SettingValida
     }, DEFAULT_SETTINGS);
 
     validateProperty(settings, 'previewNewDesign', isBoolean, DEFAULT_SETTINGS);
+    validateProperty(settings, 'previewNewestDesign', isBoolean, DEFAULT_SETTINGS);
     validateProperty(settings, 'enableForPDF', isBoolean, DEFAULT_SETTINGS);
     validateProperty(settings, 'enableForProtectedPages', isBoolean, DEFAULT_SETTINGS);
     validateProperty(settings, 'enableContextMenus', isBoolean, DEFAULT_SETTINGS);