+++ tasks/build.js
@@ -13,7 +13,7 @@ import * as reload from './reload.js';
 import codeStyle from './code-style.js';
 import zip from './zip.js';
 import {runTasks} from './task.js';
-import {log} from './utils.js';
+import {log, pathExistsSync} from './utils.js';
 import process from 'node:process';
 import {PLATFORM} from './platform.js';
 
@@ -109,6 +109,7 @@ function getParams(args) {
     };
     const platforms = {
         [PLATFORM.CHROMIUM_MV2]: false,
+        [PLATFORM.CHROMIUM_MV2_PLUS]: false,
         [PLATFORM.CHROMIUM_MV3]: false,
         [PLATFORM.FIREFOX_MV2]: false,
         [PLATFORM.THUNDERBIRD]: false,
@@ -120,6 +121,10 @@ function getParams(args) {
             allPlatforms = false;
         }
     }
+    if ((args.includes('--chrome') || args.includes('--chrome-mv2')) && args.includes('--plus')) {
+        platforms[PLATFORM.CHROMIUM_MV2] = false;
+        platforms[PLATFORM.CHROMIUM_MV2_PLUS] = true;
+    }
     if (allPlatforms) {
         Object.keys(platforms).forEach((platform) => platforms[platform] = true);
     }
@@ -128,6 +133,10 @@ function getParams(args) {
     if (platforms[PLATFORM.FIREFOX_MV3]) {
         platforms[PLATFORM.FIREFOX_MV3] = false;
         console.log('Firefox MV3 build is not supported yet');
+    }
+
+    if (!pathExistsSync('./src/plus/')) {
+        platforms[PLATFORM.CHROMIUM_MV2_PLUS] = false;
     }
 
     const versionArg = args.find((a) => a.startsWith('--version='));