+++ tasks/bundle-manifest.js
@@ -8,8 +8,9 @@ import {copyFile} from 'node:fs/promises';
 
 async function patchManifest(platform, debug, watch, test) {
     const manifest = await readJSON(absolutePath('src/manifest.json'));
-    const manifestPatch = platform === PLATFORM.CHROMIUM_MV2 ? {} : await readJSON(absolutePath(`src/manifest-${platform}.json`));
-    const patched = {...manifest, ...manifestPatch};
+    const manifestPatch = platform === PLATFORM.CHROMIUM_MV2 || platform === PLATFORM.CHROMIUM_MV2_PLUS ? {} : await readJSON(absolutePath(`src/manifest-${platform.replace('-plus', '')}.json`));
+    const manifestExtras = platform === PLATFORM.CHROMIUM_MV2_PLUS ? await readJSON(absolutePath(`src/plus/manifest.json`)) : {};
+    const patched = {...manifest, ...manifestPatch, ...manifestExtras};
     if (debug && platform === PLATFORM.CHROMIUM_MV3) {
         patched.name = 'Dark Reader MV3';
     }
@@ -30,6 +31,9 @@ async function patchManifest(platform, debug, watch, test) {
     }
     if (debug && (platform === PLATFORM.CHROMIUM_MV2 || platform === PLATFORM.CHROMIUM_MV3)) {
         patched.version_name = 'Debug';
+    }
+    if (debug && platform === PLATFORM.CHROMIUM_MV2_PLUS) {
+        patched.version_name = 'Debug Plus';
     }
     // Needed to test settings export and CSS theme export via a download
     if (test || debug) {
@@ -59,7 +63,7 @@ const bundleManifestTask = createTask(
         const chrome = changedFiles.some((file) => file.endsWith('manifest.json'));
         const platforms = {};
         for (const platform of Object.values(PLATFORM)) {
-            const changed = chrome || changedFiles.some((file) => file.endsWith(`manifest-${platform}.json`));
+            const changed = chrome || changedFiles.some((file) => file.endsWith(`manifest-${platform.replace('-plus', '')}.json`));
             platforms[platform] = changed && buildPlatforms[platform];
         }
         await manifests({platforms, debug: true, watch: true, test: false});