+++ tasks/translate.js
@@ -181,7 +181,7 @@ async function translate(text, lang) {
     })).toString();
     const response = await httpsRequest(url.toString());
     const translation = JSON.parse(response.text());
-    return translation.sentences.map((s) => s.trans).join('\n');
+    return translation.sentences.map((s) => s.trans).join('\n').replaceAll(/\n+/g, '\n');
 }
 
 const args = process.argv.slice(2);