+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# 4.9.90 (September 1, 2024)
+
+- Fixed "Aw, Snap" in Google Search (Chrome and Edge on Windows).
+
+# 4.9.89 (August 23, 2024)
+
+- Removed www part from website URLs.
+- Enabled automation and dark theme detection by default for mobile version.
+- Fixed colors for static site fixes.
+- Apply ignored image selectors for gradients.
+
+# 4.9.88 (July 11, 2024)
+
+- Fixed breaking browser panels.
+
 # 4.9.87 (June 27, 2024)
 
 - Added link to Settings page.